Renamed verificators #1102
The suite files splitted on tools and framwork subdirs. Test cases files
splitted on verifiation and case subdirs
