+++ tests/test_api/Tests/Verificators/GetAccountVerificator.py
@@ -1,4 +1,4 @@
-from Tests.CasesVerification.OutputBaseVerification import *
+from Tests.Verificators.OutputBaseVerificator import *
 
 import unittest
 import hashlib
