+++ tests/test_api/Tests/Verificators/OutputBaseVerificator.py
